Started integrating swagger
Integrated sails-swagger package which simplifies the use of swagger and does all the configurations for us.
In addition, used swagger-ui and directed the sails-swagger config (which lays in the environment setting - because I need access to the app url,
otherwise I have no access from an outside config file) to the ui url.

Use /swagger/ui to access the user friendly UI of swagger about the API.
(/swagger/doc gives you an ugly json, but if you're an external app this might be useful for programming).